Mysterious talisman
this fix isnt really a fix the item is working in game as it should but for it to do so i had to manually add the <baseAdd/> command in there. the skill id is the correct one but it does not seem to apply the stat itself. sadly i dont know how to fix skills themselves probably need java touches.